fix: rename "mockers" to "jest-mocks"
